+++ src/ClipPath.php
@@ -6,8 +6,8 @@ namespace UIAwesome\Html\Svg;
 
 use BackedEnum;
 use InvalidArgumentException;
+use UIAwesome\Html\Core\Element\BaseBlock;
 use UIAwesome\Html\Helper\Validator;
-use UIAwesome\Html\Svg\Base\BaseSvgBlockTag;
 use UIAwesome\Html\Svg\Exception\Message;
 use UIAwesome\Html\Svg\Tag\SvgBlock;
 use UIAwesome\Html\Svg\Values\{CoordinateUnits, SvgAttribute};
@@ -36,9 +36,9 @@ use UIAwesome\Html\Svg\Values\{CoordinateUnits, SvgAttribute};
  * ```
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/SVG/Element/clipPath
- * {@see BaseSvgBlockTag} for the base implementation.
+ * {@see BaseBlock} for the base block-level implementation.
  */
-final class ClipPath extends BaseSvgBlockTag
+final class ClipPath extends BaseBlock
 {
     /**
      * Sets the `clipPathUnits` attribute for the `<clipPath>` element.