+++ src/RadialGradient.php
@@ -7,13 +7,11 @@ namespace UIAwesome\Html\Svg;
 use BackedEnum;
 use InvalidArgumentException;
 use Stringable;
+use UIAwesome\Html\Core\Element\BaseBlock;
 use UIAwesome\Html\Helper\Validator;
-use UIAwesome\Html\Svg\Base\BaseSvgBlockTag;
 use UIAwesome\Html\Svg\Exception\Message;
 use UIAwesome\Html\Svg\Tag\SvgBlock;
-use UIAwesome\Html\Svg\Values\CoordinateUnits;
-use UIAwesome\Html\Svg\Values\SpreadMethod;
-use UIAwesome\Html\Svg\Values\SvgAttribute;
+use UIAwesome\Html\Svg\Values\{CoordinateUnits, SpreadMethod, SvgAttribute};
 use UnitEnum;
 
 /**
@@ -41,9 +39,9 @@ use UnitEnum;
  * ```
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/SVG/Reference/Element/radialGradient
- * {@see BaseSvgBlockTag} for the base implementation.
+ * {@see BaseBlock} for the base block-level implementation.
  */
-final class RadialGradient extends BaseSvgBlockTag
+final class RadialGradient extends BaseBlock
 {
     /**
      * Sets the SVG `cx` attribute for the element.